+++ packages/Admetica/dockerfiles/Admetica.py
@@ -1,9 +1,5 @@
 import os
-import csv
-import subprocess
-import joblib
 import torch
-import chemprop
 import numpy as np
 import logging
 from time import time