+++ packages/Bio/src/package.ts
@@ -47,11 +47,10 @@ import {BioSubstructureFilter} from './widgets/bio-substructure-filter';
 import {WebLogoViewer} from './viewers/web-logo-viewer';
 import {MonomerLibManager} from './utils/monomer-lib/lib-manager';
 import {getMonomerLibraryManagerLink, showManageLibrariesDialog, showManageLibrariesView} from './utils/monomer-lib/library-file-manager/ui';
-import {demoBio01UI} from './demo/bio01-similarity-diversity';
-import {demoBio01aUI} from './demo/bio01a-hierarchical-clustering-and-sequence-space';
+import {demoBioSimDiv} from './demo/bio01-similarity-diversity';
+import {demoSeqSpace} from './demo/bio01a-hierarchical-clustering-and-sequence-space';
 import {demoBio01bUI} from './demo/bio01b-hierarchical-clustering-and-activity-cliffs';
-import {demoBio03UI} from './demo/bio03-atomic-level';
-import {demoBio05UI} from './demo/bio05-helm-msa-sequence-space';
+import {demoToAtomicLevel} from './demo/bio03-atomic-level';
 import {checkInputColumnUI} from './utils/check-input-column';
 import {MsaWarning} from './utils/multiple-sequence-alignment';
 import {multipleSequenceAlignmentUI} from './utils/multiple-sequence-alignment-ui';
@@ -1096,10 +1095,9 @@ export function addCopyMenu(cell: DG.Cell, menu: DG.Menu): void {
 //meta.demoPath: Bioinformatics | Similarity, Diversity
 //description: Sequence similarity tracking and evaluation dataset diversity
 //meta.path: /apps/Tutorials/Demo/Bioinformatics/Similarity,%20Diversity
-//meta.isDemoScript: True
 //meta.demoSkip: GROK-14320
 export async function demoBioSimilarityDiversity(): Promise<void> {
-  await demoBio01UI();
+  await demoBioSimDiv();
 }
 
 // demoBio01a
@@ -1107,10 +1105,9 @@ export async function demoBioSimilarityDiversity(): Promise<void> {
 //meta.demoPath: Bioinformatics | Sequence Space
 //description: Exploring sequence space of Macromolecules, comparison with hierarchical clustering results
 //meta.path: /apps/Tutorials/Demo/Bioinformatics/Sequence%20Space
-//meta.isDemoScript: True
 //meta.demoSkip: GROK-14320
 export async function demoBioSequenceSpace(): Promise<void> {
-  await demoBio01aUI();
+  await demoSeqSpace();
 }
 
 // demoBio01b
@@ -1129,21 +1126,9 @@ export async function demoBioActivityCliffs(): Promise<void> {
 //meta.demoPath: Bioinformatics | Atomic Level
 //description: Atomic level structure of Macromolecules
 //meta.path: /apps/Tutorials/Demo/Bioinformatics/Atomic%20Level
-//meta.isDemoScript: True
 //meta.demoSkip: GROK-14320
 export async function demoBioAtomicLevel(): Promise<void> {
-  await demoBio03UI();
-}
-
-// demoBio05
-//name: demoBioHelmMsaSequenceSpace
-//meta.demoPath: Bioinformatics | Helm, MSA, Sequence Space
-//description: MSA and composition analysis on Helm data
-//meta.path: /apps/Tutorials/Demo/Bioinformatics/Helm,%20MSA,%20Sequence%20Space
-//meta.isDemoScript: True
-//meta.demoSkip: GROK-14320
-export async function demoBioHelmMsaSequenceSpace(): Promise<void> {
-  await demoBio05UI();
+  await demoToAtomicLevel();
 }
 
 //name: SDF to JSON Library