+++ packages/Chembl/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Chembl changelog
 
+## 1.3.0 (2025-03-25)
+
+* Updated chemical databases demo
+
 ## 1.2.3 (2025-03-17)
 
 * Showing update indicator in context panel when waiting for request